GVT-2904: Split publication-candidate-layer into layer and utils (will probably need to be split even further), add id to map tools, unify duplicate code, remove debug loggings and turn == into === in some places among other things
