+++ ui/src/map/tools/highlight-tool.ts
@@ -6,6 +6,7 @@ import { getDefaultHitArea, searchItemsFromLayers } from 'map/tools/tool-utils';
 
 let currentItemsCompare = '';
 export const highlightTool: MapTool = {
+    id: 'highlight',
     activate: (map: OlMap, layers: MapLayer[], options: MapToolActivateOptions) => {
         const debouncedMoveHandlerHighlight = debounce(
             ({ coordinate }) => {
