+++ ui/src/map/tools/tool-model.ts
@@ -27,8 +27,10 @@ export type MapTool = {
     activate: (map: OlMap, layers: MapLayer[], options: MapToolActivateOptions) => DeactivateToolFn;
     customCursor?: CssType.Property.Cursor;
     component?: React.FC<MapToolProps>;
+    id: string;
 };
 
 export type SelectableMapTool = MapTool & {
     component: React.FC<MapToolProps>;
+    id: string;
 };
