+++ ui/src/model/geometry.ts
@@ -181,6 +181,11 @@ export const rangeContainsExclusive = (range: Range<number>, value: number): boo
 export const rangeContainsInclusive = (range: Range<number>, value: number): boolean =>
     range.min <= value && range.max >= value;
 
+export const mergeRanges = (range1: Range<number>, range2: Range<number>): Range<number> => ({
+    min: Math.min(range1.min, range2.min),
+    max: Math.max(range1.max, range2.max),
+});
+
 export const centerForBoundingBox = (bbox: BoundingBox): Point => ({
     x: (bbox.x.min + bbox.x.max) / 2,
     y: (bbox.y.min + bbox.y.max) / 2,
